requirements.txt
astroid==1.4.4
colorama==0.3.6
Django==1.9.1
djangorestframework==3.3.2
ecdsa==0.13
Fabric==1.10.2
lazy-object-proxy==1.2.1
paramiko==1.16.0
pycrypto==2.6.1
pylint==1.5.4
pylint-django==0.7.1
pylint-plugin-utils==0.2.3
six==1.10.0
wheel==0.24.0
wrapt==1.10.6
